requirements.txt
# Python dependencies for web application
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Optional: For generating test sequences
Pillow==10.0.0

# Development dependencies
pytest==7.4.0
pytest-cov==4.1.0

# Note: FFmpeg must be installed separately on the system
